rodeo/requirements.txt
ipykernel == 4.0.3
rodeo == 0.4.3
flask-cdn == 1.2.1